Refactor to remove Query and use only GenericQuery

+++ api/src/main/java/api/mappers/TreeNeighbourhoodPairMapper.java
@@ -4,12 +4,12 @@ import api.TreeRecord;
 import com.hazelcast.mapreduce.Context;
 import com.hazelcast.mapreduce.Mapper;
 
-public class TreeNeighbourhoodPairMapper implements Mapper<String, Long, Long, String> {
+public class TreeNeighbourhoodPairMapper implements Mapper<String, Integer, Integer, String> {
 
     @Override
-    public void map(String neighbourhood, Long treeCount, Context<Long, String> context) {
+    public void map(String neighbourhood, Integer treeCount, Context<Integer, String> context) {
         // Round down to the nearest thousand
-        long roundedTreeCount = treeCount - (treeCount % 1000);
+        int roundedTreeCount = treeCount - (treeCount % 1000);
         context.emit(roundedTreeCount, neighbourhood);
     }
 }
